% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{build_filter}
\alias{build_filter}
\title{Get participants}
\usage{
build_filter(var, vals)
}
\description{
Get participants
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante-example-dataset:bm7r", version = "current"))
participants <- get_participants(dataset_spec)
}
}
