% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{collect_users}
\alias{collect_users}
\title{Collect user data}
\usage{
collect_users(dataset_data)
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
longer description
}
