% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine tables into a single megatable}
\usage{
combine_datasets(dataset_tables)
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
longer description
}
