% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-ModelRecord.R
\name{count_items}
\alias{count_items}
\title{count number of items with non-NA responses for each run}
\usage{
count_items(mod_rec)
}
\arguments{
\item{mod_rec}{ModelRecord object}
}
\description{
count number of items with non-NA responses for each run
}
