% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_caching_functions.R
\docType{data}
\name{dataset_cache_dir}
\alias{dataset_cache_dir}
\title{Functions to cache and retrieve Redivis Datasets}
\format{
An object of class \code{character} of length 1.
}
\usage{
dataset_cache_dir
}
\description{
Functions to cache and retrieve Redivis Datasets
}
\keyword{datasets}
