% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_corpus_items}
\alias{fetch_corpus_items}
\title{Get metadata for corpus items}
\usage{
fetch_corpus_items()
}
\description{
Get metadata for corpus items
}
\examples{
\dontrun{
corpus_items <- fetch_corpus_items()
}
}
\keyword{internal}
