% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_item_mapping_fields}
\alias{fetch_item_mapping_fields}
\title{Get item UIDs mapped by fields}
\usage{
fetch_item_mapping_fields()
}
\description{
Get item UIDs mapped by fields
}
\examples{
\dontrun{
mapping_fields <- fetch_item_mapping_fields()
}
}
\keyword{internal}
