% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_item_mapping_trial}
\alias{fetch_item_mapping_trial}
\title{Get item UIDs mapped by trial_id}
\usage{
fetch_item_mapping_trial()
}
\description{
Get item UIDs mapped by trial_id
}
\examples{
\dontrun{
mapping_trial <- fetch_item_mapping_trial()
}
}
