% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_mapping_items}
\alias{fetch_mapping_items}
\title{Get metadata for mapping items}
\usage{
fetch_mapping_items()
}
\description{
Get metadata for mapping items
}
\examples{
\dontrun{
mapping_items <- fetch_mapping_items()
}
}
\keyword{internal}
