% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{fetch_registry_dir}
\alias{fetch_registry_dir}
\title{Fetch model registry file index table}
\usage{
fetch_registry_dir(version = "current")
}
\arguments{
\item{version}{Version tag of the \code{levante_metadata_scoring} dataset.}
}
\description{
Fetch model registry file index table
}
