% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{fetch_registry_table}
\alias{fetch_registry_table}
\title{Fetch model registry file index table}
\usage{
fetch_registry_table(version = "current")
}
\description{
Fetch model registry file index table
}
\keyword{internal}
