% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{fetch_scoring_table}
\alias{fetch_scoring_table}
\title{Fetch scoring specification table}
\usage{
fetch_scoring_table(version = "current")
}
\arguments{
\item{version}{Version tag of the \code{levante_metadata_scoring} dataset.}
}
\description{
Fetch scoring specification table
}
