% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_survey_items}
\alias{fetch_survey_items}
\title{Get metadata for survey items}
\usage{
fetch_survey_items()
}
\description{
Get metadata for survey items
}
\examples{
\dontrun{
survey_items <- fetch_survey_items()
}
}
\keyword{internal}
