% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_trial_items}
\alias{fetch_trial_items}
\title{Get metadata for trial items}
\usage{
fetch_trial_items()
}
\description{
Get metadata for trial items
}
\examples{
\dontrun{
trial_items <- fetch_trial_items()
}
}
\keyword{internal}
