% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-fit.R
\name{fit_bylanguage_lang}
\alias{fit_bylanguage_lang}
\title{wrapper: given language, fit by_language model for each task}
\usage{
fit_bylanguage_lang(task_data, models, priors, lang, registry_dir)
}
\arguments{
\item{task_data}{dataframe with columns item_task, the value of subset_var,
and data, where data is trial-level data}

\item{models}{dataframe of models to fit, specified with columns nfact
(number of factors), itemtype (Rasch/2PL), invariance
(configural/metric/scalar)}

\item{priors}{list of priors, where names are parameter names (e.g. d, a1)
and values are vectors of length 3 (priorType, val1, val2)}

\item{lang}{string indicating language}

\item{registry_dir}{string indicating directory in which to save models}
}
\description{
wrapper: given language, fit by_language model for each task
}
