% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{fix_table_types}
\alias{fix_table_types}
\title{Fix some stuff in tables}
\usage{
fix_table_types(table_data)
}
\value{
A Levante specific function to clean up some known potential data issues.
}
\description{
There can be some anomalies in LEVANTE data stored in Redivis. This function
will clean up some of the most common
}
