% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-helpers.R
\name{generate_model_str_numeric}
\alias{generate_model_str_numeric}
\title{generates mirt model strings}
\usage{
generate_model_str_numeric(
  df,
  df_prepped,
  item_type,
  f,
  priors = NULL,
  item_sep = "-"
)
}
\arguments{
\item{df}{trial data}

\item{df_prepped}{trial data in wide mirt shape (runs x items)}

\item{item_type}{"Rasch", "2PL", etc}

\item{f}{number of factors}

\item{priors}{list of priors}

\item{item_sep}{string to put in between item_uid and instance index (defaults to "-")}
}
\value{
string that can be passed to \code{mirt.model()} as \code{input}
}
\description{
generates mirt model strings
}
