% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{get_datasets}
\alias{get_datasets}
\title{Get redivis datasets}
\usage{
get_datasets(dataset_names, org_name = "levante", tables = NULL)
}
\value{
A list of one or more datasets from a specific organization in the Redivis repository. In our case that is normally "levante"
}
\description{
Retrieve LEVANTE project datasets from Redivis and prepare them for further analysis.
}
