% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{get_model_record}
\alias{get_model_record}
\title{get the model record indexed in registry_dir for a given scoring specification}
\usage{
get_model_record(spec, registry_dir)
}
\arguments{
\item{spec}{list with entries item_task, model_set, subset, itemtype, nfact, invariance}

\item{registry_dir}{tibble returned by registry_dir()}
}
\value{
ModelRecord object
}
\description{
get the model record indexed in registry_dir for a given scoring specification
}
\keyword{internal}
