% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.r
\name{get_model_spec}
\alias{get_model_spec}
\title{get the scoring specification in scoring_table for a given task + dataset}
\usage{
get_model_spec(scoring_table, score_task, score_dataset)
}
\description{
get the scoring specification in scoring_table for a given task + dataset
}
