% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-raw_data.R
\name{get_raw_table}
\alias{get_raw_table}
\title{Get raw table}
\usage{
get_raw_table(table_name, data_source, data_source_version = "current")
}
\arguments{
\item{table_name}{String indicating name of table to get}

\item{data_source}{String indicating which Redivis dataset to get}

\item{data_source_version}{String indicating which version of \code{data_source} to get}
}
\value{
A data frame corresponding to the specified Redivis table.
}
\description{
\code{get_raw_table()} returns any table from Redivis as a data frame. See the \href{https://levante-framework.github.io/rlevante/index.html}{rlevante documentation} for more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
runs <- get_raw_table(table_name = "runs", data_source = "levante_data_example_raw:bm7r")
}
}
