% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{get_registry_file}
\alias{get_registry_file}
\title{fetch file from model registry}
\usage{
get_registry_file(mod_filename, registry_dir)
}
\arguments{
\item{mod_filename}{Filename (string) of model record to fetch.}

\item{registry_dir}{Model registry directory as returned by \code{fetch_registry_dir()}.}
}
\description{
fetch file from model registry
}
