% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_survey_items}
\alias{get_survey_items}
\title{Get metadata for survey items}
\usage{
get_survey_items()
}
\description{
Get metadata for survey items
}
\examples{
\dontrun{
survey_items <- get_survey_items()
}
}
