% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed_data.R
\name{get_surveys}
\alias{get_surveys}
\title{Get surveys}
\usage{
get_surveys(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row is a survey item response.
}
\description{
\code{get_surveys()} returns information about survey responses as a data frame. See the \href{https://levante-framework.github.io/rlevante/index.html}{rlevante documentation} for more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
surveys <- get_surveys(dataset = "levante_data_example:d0rt", version = "current")
}
}
