% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levantemodels-package.R
\docType{package}
\name{levantemodels-package}
\alias{levantemodels}
\alias{levantemodels-package}
\title{levantemodels: LEVANTE Data Processing and Modeling}
\description{
Internal tooling for processing LEVANTE data and fitting and serving its psychometric (IRT) scoring models. Holds the trial recoding, scoring pipeline, model registry, and ModelRecord machinery. For loading LEVANTE data, see the levante package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://levante-framework.github.io/levantemodels/}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com} (\href{https://orcid.org/0000-0001-9039-3220}{ORCID})

Authors:
\itemize{
  \item Michael Frank \email{mcfrank@stanford.edu} (\href{https://orcid.org/0000-0002-7551-4378}{ORCID})
}

Other contributors:
\itemize{
  \item Theresa Cheng \email{twcheng@stanford.edu} (\href{https://orcid.org/0000-0002-8557-565X}{ORCID}) [contributor]
}

}
\keyword{internal}
