% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{mirt_aic}
\alias{mirt_aic}
\title{get AIC of fitted mirt model}
\usage{
mirt_aic(mod)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}
}
\description{
get AIC of fitted mirt model
}
