% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{mirt_bic}
\alias{mirt_bic}
\title{get BIC of fitted mirt model}
\usage{
mirt_bic(mod)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}
}
\description{
get BIC of fitted mirt model
}
