% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{mirt_coefs}
\alias{mirt_coefs}
\title{get item parameters of fitted mirt model}
\usage{
mirt_coefs(mod, IRTpars = TRUE)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}

\item{IRTpars}{logical; convert slope intercept parameters into traditional IRT parameters?}
}
\description{
get item parameters of fitted mirt model
}
