% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{mirt_scores}
\alias{mirt_scores}
\title{get participant scores of fitted mirt model}
\usage{
mirt_scores(mod, df, df_prepped)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}

\item{df}{trial data}

\item{df_prepped}{trial data in wide mirt shape (runs x items)}
}
\description{
get participant scores of fitted mirt model
}
