% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-ModelRecord.R
\name{model_from_record}
\alias{model_from_record}
\title{given ModelRecord object, instantiate mirt model object}
\usage{
model_from_record(mod_rec)
}
\arguments{
\item{mod_rec}{ModelRecord object}
}
\description{
given ModelRecord object, instantiate mirt model object
}
