% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{model_spec_filename}
\alias{model_spec_filename}
\title{convert model spec to filename}
\usage{
model_spec_filename(spec)
}
\arguments{
\item{spec}{Model specification (list with names item_task, model_set, subset, itemtype, nfact, invariance).}
}
\description{
convert model spec to filename
}
