% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-ModelRecord.R
\name{modelrecord}
\alias{modelrecord}
\alias{extract_scores}
\alias{show,ModelRecord-method}
\alias{AIC,ModelRecord-method}
\alias{BIC,ModelRecord-method}
\alias{logLik,ModelRecord-method}
\alias{model_class}
\alias{model_vals}
\alias{items}
\alias{scores}
\alias{tabdata}
\title{constructor helper}
\usage{
modelrecord(mod, row_ids)

extract_scores(mod, row_ids)

\S4method{show}{ModelRecord}(object)

\S4method{AIC}{ModelRecord}(object)

\S4method{BIC}{ModelRecord}(object)

\S4method{logLik}{ModelRecord}(object)

model_class(object)

model_vals(object)

items(object)

scores(object)

tabdata(object)
}
\arguments{
\item{mod}{mirt model object}

\item{row_ids}{character vector}

\item{object}{ModelRecord object}
}
\description{
constructor helper

helper to extract and tidy scores

method for show generic

method for AIC generic

method for BIC generic

method for logLik generic
}
