% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{multigroup_coefs}
\alias{multigroup_coefs}
\title{get coefficients for multigroup model}
\usage{
multigroup_coefs(mod)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}
}
\description{
get coefficients for multigroup model
}
