% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{multigroup_extract_groups}
\alias{multigroup_extract_groups}
\title{get groups out of multigroup model}
\usage{
multigroup_extract_groups(mod)
}
\arguments{
\item{mod}{fit mirt model (SingleGroupClass or MultiGroupClass object)}
}
\description{
get groups out of multigroup model
}
