% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-extract.R
\name{multigroup_itemfit}
\alias{multigroup_itemfit}
\title{get item fits out of multigroup model}
\usage{
multigroup_itemfit(submods, fit_stats)
}
\arguments{
\item{submods}{list of SingleGroupClass mirt models}

\item{fit_stats}{character vector indicating which fit statistics should be computed}
}
\description{
get item fits out of multigroup model
}
