% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notebook_dataset}
\alias{notebook_dataset}
\title{Determine dataset of current notebook on Redivis}
\usage{
notebook_dataset(redivis)
}
\arguments{
\item{redivis}{Redivis reference object.}
}
\description{
Determine dataset of current notebook on Redivis
}
\keyword{internal}
