% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-runs.R
\name{process_runs}
\alias{process_runs}
\title{Process run data}
\usage{
process_runs(
  dataset_spec,
  remove_incomplete_runs = TRUE,
  remove_invalid_runs = TRUE,
  max_results = NULL
)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}

\item{remove_incomplete_runs}{Boolean indicating whether to drop runs that
were marked as incomplete (defaults to TRUE).}

\item{remove_invalid_runs}{Boolean indicating whether to drop runs that were
marked as invalid (defaults to TRUE).}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\description{
Process run data
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante_data_example:d0rt", version = "current"))
runs <- process_runs(dataset_spec)
}
}
\keyword{internal}
