% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-trials.R
\name{process_trials_prelim}
\alias{process_trials_prelim}
\title{Process trial data}
\usage{
process_trials_prelim(
  dataset_spec,
  remove_incomplete_runs = TRUE,
  remove_invalid_runs = TRUE,
  remove_invalid_trials = FALSE,
  tasks = NULL,
  participants = NULL,
  max_results = NULL
)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}

\item{remove_incomplete_runs}{Boolean indicating whether to drop runs that
were marked as incomplete (defaults to TRUE).}

\item{remove_invalid_runs}{Boolean indicating whether to drop runs that were
marked as invalid (defaults to TRUE).}

\item{remove_invalid_trials}{Boolean indicating whether to drop trials that
were marked as invalid (defaults to FALSE).}

\item{tasks}{Character vector of tasks to include.}

\item{participants}{(Optional) Data frame that includes the columns "dataset"
and "user_id", if supplied trial data will be filtered to only those user
IDs.}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\description{
Process trial data
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante_data_example:d0rt", version = "current"))
trials_prelim <- process_trials_prelim(dataset_spec)
}
}
\keyword{internal}
