% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_prep_helpers.R
\name{recode_hf}
\alias{recode_hf}
\title{recode correctness + reclassify items for HF}
\usage{
recode_hf(df)
}
\arguments{
\item{df}{trial data}
}
\description{
recode correctness + reclassify items for HF
}
\keyword{internal}
