% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_prep_helpers.R
\name{recode_slider}
\alias{recode_slider}
\title{recode correctness for slider}
\usage{
recode_slider(df, threshold)
}
\arguments{
\item{df}{trial data}

\item{threshold}{max normalized distance from slider target}
}
\description{
recode correctness for slider
}
