% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_prep_helpers.R
\name{recode_wrong_items}
\alias{recode_wrong_items}
\title{recode correctness for items with wrong answers}
\usage{
recode_wrong_items(df, wrong_items)
}
\arguments{
\item{df}{trial data}

\item{wrong_items}{tibble with columns item_uid and answer_fixed}
}
\description{
recode correctness for items with wrong answers
}
\keyword{internal}
