% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-irt_helpers.R
\name{remove_no_var_items}
\alias{remove_no_var_items}
\title{remove items with no variance}
\usage{
remove_no_var_items(df)
}
\arguments{
\item{df}{trial data}
}
\description{
remove items with no variance
}
\keyword{internal}
