% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-irt_helpers.R
\name{remove_no_var_items_bygroup}
\alias{remove_no_var_items_bygroup}
\title{remove items with no invariance in any single group}
\usage{
remove_no_var_items_bygroup(df, item_n_min = 1)
}
\arguments{
\item{df}{trial data}

\item{item_n_min}{minimum number of times an item needs to have been seen to be retained}
}
\description{
remove items with no invariance in any single group
}
\keyword{internal}
