% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-irt_helpers.R
\name{remove_nonshared_items}
\alias{remove_nonshared_items}
\title{remove items that aren't shared across all groups}
\usage{
remove_nonshared_items(df)
}
\arguments{
\item{df}{trial data}
}
\description{
remove items that aren't shared across all groups
}
\keyword{internal}
