% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{remove_practice_trials}
\alias{remove_practice_trials}
\title{Remove practice trials from trial data}
\usage{
remove_practice_trials(trials)
}
\arguments{
\item{trials}{Tibble of trial data as returned by combine_datasets().}
}
\value{
Trials tibble with any non-test trials removed (instructions,
  practice, training).
}
\description{
Remove practice trials from trial data
}
\examples{
ds <- get_datasets("ca_western_pilot", tables = "trials", max_results = 5)
ds_combined <- combine_datasets(ds)
ds_trials <- remove_practice_trials(ds_combined$trials)
}
