% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-rescore.R
\name{rescore}
\alias{rescore}
\title{given trial data and model record, score data from corresponding model}
\usage{
rescore(trial_data_task, mod_spec, mod_rec)
}
\description{
given trial data and model record, score data from corresponding model
}
