% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(task, dataset, trials, runs, scoring_table, registry_dir)
}
\arguments{
\item{task}{task}

\item{dataset}{dataset}

\item{trials}{trial data from one task and one dataset}

\item{runs}{run data from one task and one dataset}

\item{scoring_table}{tibble returned by fetch_scoring_table()}

\item{registry_dir}{redivis directory returned by fetch_registry_dir()}
}
\description{
score
}
\keyword{internal}
