% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-score.R
\name{score_pa}
\alias{score_pa}
\title{scores for PA}
\usage{
score_pa(trial_data_task, dataset)
}
\arguments{
\item{trial_data_task}{trial data from one task and one dataset}

\item{dataset}{dataset}
}
\description{
scores for PA
}
