% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-score.R
\name{score_sre}
\alias{score_sre}
\title{scores for SRE}
\usage{
score_sre(trial_data_task, dataset)
}
\arguments{
\item{trial_data_task}{trial data from one task and one dataset}

\item{dataset}{dataset}
}
\description{
scores for SRE
}
