% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-irt_helpers.R
\name{to_mirt_shape_grouped}
\alias{to_mirt_shape_grouped}
\title{format data for mirt}
\usage{
to_mirt_shape_grouped(df)
}
\arguments{
\item{df}{trial data}
}
\description{
format data for mirt
}
\keyword{internal}
