% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine tables into a single megatable}
\usage{
combine_datasets(dataset_tables)
}
\value{
For the case where the data you're using is in multiple tables, this
 function will combine them into one single large table if needed.
}
\description{
longer description
}
