% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-irt_helpers.R
\name{dedupe_items}
\alias{dedupe_items}
\title{add identifiers for each instance of each item}
\usage{
dedupe_items(df, item_sep = "-")
}
\arguments{
\item{df}{trial data}

\item{item_sep}{string to put in between item_uid and instance index (defaults to "-")}
}
\description{
add identifiers for each instance of each item
}
