% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{fetch_registry_dir}
\alias{fetch_registry_dir}
\title{Fetch model registry file index table}
\usage{
fetch_registry_dir(version = "current")
}
\description{
Fetch model registry file index table
}
\keyword{internal}
