% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{fetch_scoring_table}
\alias{fetch_scoring_table}
\title{Fetch scoring specification table}
\usage{
fetch_scoring_table(version = "current")
}
\description{
Fetch scoring specification table
}
\keyword{internal}
