% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{fix_table_types}
\alias{fix_table_types}
\title{Fix some stuff in tables}
\usage{
fix_table_types(table_data)
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
longer description
}
