% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{get_datasets}
\alias{get_datasets}
\title{Get redivis datasets}
\usage{
get_datasets(dataset_names, org_name = "levante", tables = NULL)
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
longer description
}
