% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_functions.R
\name{get_datasets}
\alias{get_datasets}
\title{Get data from one or more datasets}
\usage{
get_datasets(
  dataset_names,
  org_name = "levante",
  tables = NULL,
  max_results = NULL
)
}
\arguments{
\item{dataset_names}{Character vector of dataset names to fetch.}

\item{org_name}{Redivis organization name (defaults to "levante")}

\item{tables}{Character vector of table names to fetch (default to all
tables).}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\value{
Nested list of datasets, each of which is a list whose names are
  table names and values are tibbles of data for that dataset + table.
}
\description{
Get data from one or more datasets
}
\examples{
dataset_names <- list(list(name = "ca_western_pilot:97mt", version = "v3_1"))
get_datasets(dataset_names, tables = "users", max_results = 5)
}
