% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{get_item_parameters}
\alias{get_item_parameters}
\title{Get item parameters}
\usage{
get_item_parameters(version = "current")
}
\arguments{
\item{version}{Version of the Redivis scoring metadata dataset.}
}
\value{
A data frame where each row is an item parameter estimate. See our \href{https://researcher.levante-network.org/measures/scoring-and-psychometrics}{Scoring and Psychometrics page} to learn how to interpret these values.
}
\description{
\code{get_item_parameters()} returns the IRT item parameters used in LEVANTE scoring as a data frame. See the \href{https://levante-framework.github.io/rlevante/index.html}{rlevante documentation} for more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
item_parameters <- get_item_parameters()
}
}
