% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.r
\name{get_model_record}
\alias{get_model_record}
\title{get the model record indexed in registry_table for a given scoring specification}
\usage{
get_model_record(spec, registry_table)
}
\description{
get the model record indexed in registry_table for a given scoring specification
}
