% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.r
\name{get_model_spec}
\alias{get_model_spec}
\title{get the scoring specification in scoring_table for a given task + dataset}
\usage{
get_model_spec(score_task, score_dataset, scoring_table)
}
\arguments{
\item{score_task}{string indicating task}

\item{score_dataset}{string indicating dataset}

\item{scoring_table}{tibble returned by get_scoring_table()}
}
\value{
list with entries item_task, dataset, model_set, subset, itemtype, nfact, invariance
}
\description{
get the scoring specification in scoring_table for a given task + dataset
}
