% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed_data.R
\name{get_participants}
\alias{get_participants}
\title{Get participants}
\usage{
get_participants(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\description{
Get participants
}
\examples{
\dontrun{
participants <- get_participants(dataset = "levante-data-example:bm7r", version = "current")
}
}
