% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_participants}
\alias{get_participants}
\title{Get participants}
\usage{
get_participants(dataset_spec, max_results = NULL)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get participants
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante-example-dataset:bm7r", version = "current"))
participants <- get_participants(dataset_spec)
}
}
