% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-raw_data.R
\name{get_raw_table}
\alias{get_raw_table}
\title{Get raw table}
\usage{
get_raw_table(table_name, data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\description{
Get raw table
}
\examples{
\dontrun{
runs <- get_raw_table(table_name = "runs", dataset = "levante-data-example:bm7r", version = "current")
}
}
