% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-registry.R
\name{get_registry_file}
\alias{get_registry_file}
\title{fetch file from model registry}
\usage{
get_registry_file(mod_filename, registry_table)
}
\description{
fetch file from model registry
}
\keyword{internal}
