% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_scores}
\alias{get_scores}
\title{Get score data}
\usage{
get_scores(dataset_spec)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}
}
\description{
Get score data
}
\examples{
dataset_spec <- list(list(name = "levante-example-dataset:bm7r", version = "current"))
scores <- get_scores(dataset_spec)
}
