% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed_data.R
\name{get_scores}
\alias{get_scores}
\title{Get scores}
\usage{
get_scores(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\description{
Get scores
}
\examples{
\dontrun{
scores <- get_scores(dataset = "levante_data_example:d0rt", version = "current")
}
}
