% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed_data.R
\name{get_surveys}
\alias{get_surveys}
\title{Get surveys}
\usage{
get_surveys(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\description{
Get surveys
}
\examples{
\dontrun{
surveys <- get_surveys(dataset = "levante_data_example:d0rt", version = "current")
}
}
