% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_surveys}
\alias{get_surveys}
\title{Get survey data}
\usage{
get_surveys(
  dataset_spec,
  survey_types = c("caregiver", "student", "teacher"),
  remove_incomplete_surveys = FALSE,
  max_results = NULL
)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}

\item{survey_types}{Character vector of survey types to include (caregiver,
student, teacher).}

\item{remove_incomplete_surveys}{Boolean indicating whether to drop surveys
that were marked as incomplete (defaults to FALSE).}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get survey data
}
