% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed_data.R
\name{get_trials}
\alias{get_trials}
\title{Get trials}
\usage{
get_trials(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\description{
Get trials
}
\examples{
\dontrun{
trials <- get_trials(dataset = "levante_data_example:d0rt", version = "current")
}
}
