% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_trials_raw}
\alias{get_trials_raw}
\title{Get raw trial data}
\usage{
get_trials_raw(dataset_spec)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}
}
\description{
Get raw trial data
}
