% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{link_surveys}
\alias{link_surveys}
\title{Link survey data to participant data}
\usage{
link_surveys(surveys, participants)
}
\arguments{
\item{surveys}{Survey data as returned by `get_surveys()`.}

\item{participants}{Participant data as returned by `get_participants()`.}
}
\description{
Link survey data to participant data
}
\examples{
dataset_spec <- list(list(name = "levante-example-dataset:bm7r", version = "current"))
surveys <- get_surveys(dataset_spec)
participants <- get_participants(dataset_spec)
survey_data <- surveys |>
  link_surveys(participants)
}
