% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-surveys.R
\name{link_surveys}
\alias{link_surveys}
\title{Link survey data to participant data}
\usage{
link_surveys(surveys, participants)
}
\arguments{
\item{surveys}{Survey data as returned by `process_surveys()`.}

\item{participants}{Participant data as returned by `process_participants()`.}
}
\description{
Link survey data to participant data
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante_data_example:d0rt", version = "current"))
surveys <- process_surveys(dataset_spec)
participants <- process_participants(dataset_spec)
survey_data <- surveys |> link_surveys(participants)
}
}
\keyword{internal}
