% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-participants.R
\name{process_participants}
\alias{process_participants}
\title{Process participants}
\usage{
process_participants(dataset_spec, max_results = NULL)
}
\arguments{
\item{dataset_spec}{List of dataset names and versions to retrieve.}

\item{max_results}{Max number of records to load for each table (defaults to
entire table).}
}
\description{
Process participants
}
\examples{
\dontrun{
dataset_spec <- list(list(name = "levante-example-dataset:bm7r", version = "current"))
participants <- process_participants(dataset_spec)
}
}
\keyword{internal}
