% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_prep_helpers.R
\name{recode_trials}
\alias{recode_trials}
\title{recode correctness and/or items for several tasks}
\usage{
recode_trials(df, slider_threshold = 0.15)
}
\arguments{
\item{df}{trial data}

\item{slider_threshold}{max normalized distance from slider target}
}
\description{
recode correctness and/or items for several tasks
}
\keyword{internal}
