% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlevante-package.R
\docType{package}
\name{rlevante-package}
\alias{rlevante}
\alias{rlevante-package}
\title{rlevante: Download LEVANTE Data}
\description{
Download data from LEVANTE, a global project for studying child development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://levante-framework.github.io/rlevante/}
}

}
\keyword{internal}
