% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(task, dataset, trials, runs, scoring_table, registry_table)
}
\arguments{
\item{task}{task}

\item{dataset}{dataset}

\item{trials}{trial data from one task and one dataset}

\item{runs}{run data from one task and one dataset}

\item{scoring_table}{table returned by get_scoring_table}

\item{registry_table}{table returned by get_registry_table}
}
\description{
score
}
